fix(ark-cli): use `ark-rust_hot_fix` branch of `arklib`
The branch `ark-rust_hot_fix` contains a hot fix where we use the same commit specified when writing `ark-cli` but with pinning `image` crate to the correct version so that `ark-rust` can compile

This is just a hot fix. `arklib` shouldn't be even a dependency of any `ark-rust` crate. See https://github.com/ARK-Builders/ark-rust/issues/15 for more information

Signed-off-by: Tarek <tareknaser360@gmail.com>