feat(ci): add a github workflow for binary releases
We're holding off on adding a macOS ARM check to our build workflow because it involves expenses. Testing it for every PR or commit might not be worth it, so we'll include it in the release workflow instead. Since we haven't set up a release workflow yet, we'll treat it as a reminder to run it when we do release builds.

Signed-off-by: Tarek <tareknaser360@gmail.com>